chore: use JSX in templates
