+++ frontend/package.json
@@ -23,6 +23,7 @@
     "@algolia/transporter": "4.8.4",
     "@babel/core": "7.12.10",
     "@babel/preset-env": "7.12.11",
+    "@types/react": "17.0.1",
     "algoliasearch": "4.8.4",
     "babel-loader": "8.2.2",
     "clean-webpack-plugin": "3.0.0",
