+++ frontend/tsconfig.json
@@ -4,7 +4,8 @@
     "lib": ["es2020", "DOM"],
     "rootDir": "src",
     /* Use a fake outDir */
-    "outDir": "dist-webpack-never-used"
+    "outDir": "dist-webpack-never-used",
+    "jsx": "react"
   },
   "include": ["src/**/*"]
 }
